Simplify usage and API
